Updated predict
Using the predict keyword didnt work as expected. Instead of analysing the given image, it used the all of the camvid test images. With the proposed changes main.py loads the image, resizes it to the given size and gives it to tensorflow. 
Also i had to load the class_dict from the camvid dataset. Maybe this needs to be generalized or put somewhere else.